requirements.txt
# Audio opname (Windows WASAPI loopback)
pyaudiowpatch>=0.2.12

# Spraak-naar-tekst
openai-whisper>=20231117

# Numerieke verwerking (audio mixen/resampling)
numpy>=1.24

# HTTP requests naar Ollama
requests>=2.31

# GPU-ondersteuning (installeer via aparte pip index, zie README)
# torch>=2.1  →  pip install torch --index-url https://download.pytorch.org/whl/cu121
